pyproject.toml
[project]
name = "streaming"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "kafka-python>=2.3.0",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]
